refactor: split oversized Python files under 300-line limit
- manifest.py (358→284): Extract detect_language, list_source_files,
  _walk_source_files to analysis/_detection.py. Remove dead
  _detect_category function and unused json import.
- runner.py (321→297): Extract DimensionEntry, DimensionsConfig,
  load_universal_dimensions to analysis/_dimensions.py.
- cli.py (320→258): Extract _add_evaluate_args, build_parser to
  cli_parser.py.
- runs.py (333→270): Extract _make_caching_fetcher, RunLookupCache,
  _get_previous_run_for_dimension to report_parser/_run_lookup.py.

All re-exports preserved for backward compatibility.
