+++ src/quodeq/cli.py
@@ -11,6 +11,7 @@ import uuid
 from pathlib import Path
 from typing import Callable
 
+from quodeq.cli_parser import build_parser as build_parser  # re-export
 from quodeq.config.paths import default_paths, load_env_file
 from quodeq.dashboard.cli import main as dashboard_main
 from quodeq.analysis.subprocess import AnalysisError
@@ -18,13 +19,9 @@ from quodeq.analysis.runner import AnalysisOptions, EvaluationError, RunConfig, 
 from quodeq.core.scoring.report import run_full
 from quodeq.shared.project_resolver import ProjectIdentity, resolve_project_uuid
 from quodeq.shared.repo_handler import prepare_repository
-from quodeq.shared.utils import get_evaluations_dir, is_repo_url, project_name_from_repo, write_text
+from quodeq.shared.utils import is_repo_url, project_name_from_repo, write_text
 from quodeq.shared.validation import validate_path_segment
 
-
-_DEFAULT_N_SUBAGENTS = 5
-_MODE_NUMERICAL = "numerical"
-_MODE_GRADES = "grades"
 _ENV_MAX_TURNS = "QUODEQ_MAX_TURNS"
 _ENV_MAX_DURATION = "QUODEQ_MAX_DURATION"
 
@@ -43,65 +40,6 @@ def _env_int(var: str, default: int | None, env: dict[str, str] | None = None) -
 def _subagent_model(env: dict[str, str] | None = None) -> str | None:
     """Return the subagent model override from the environment, or None."""
     return (env or os.environ).get("SUBAGENT_MODEL") or None
-
-
-def _add_evaluate_args(parser: argparse.ArgumentParser) -> None:
-    """Register arguments for the evaluate subcommand."""
-    parser.add_argument("repo", help="Path or URL to the repository")
-    parser.add_argument(
-        "-l", "--language", default=None, help="Language (overrides auto-detection)"
-    )
-    parser.add_argument(
-        "-o", "--output", default=get_evaluations_dir(), help="Reports output directory"
-    )
-    parser.add_argument(
-        "-m", "--mode", default=_MODE_NUMERICAL,
-        choices=[_MODE_NUMERICAL, _MODE_GRADES], help="Scoring mode",
-    )
-    parser.add_argument(
-        "--no-prescan", action="store_true", help="Skip source-file counting"
-    )
-    parser.add_argument(
-        "-d", "--dimensions", default=None,
-        help="Comma-separated dimensions to evaluate (default: all)",
-    )
-    parser.add_argument(
-        "--evidence-only", action="store_true",
-        help="Produce evidence JSON only (skip scoring)",
-    )
-    parser.add_argument(
-        "--max-turns", type=int, default=None,
-        help="Max AI conversation turns per dimension (default: 200)",
-    )
-    parser.add_argument(
-        "--max-duration", type=int, default=None,
-        help="Max seconds per dimension before terminating (default: 1800)",
-    )
-    parser.add_argument(
-        "--n-subagents", type=int, default=_DEFAULT_N_SUBAGENTS,
-        help="Number of parallel subagents per dimension (default: %(default)s)",
-    )
-    parser.add_argument(
-        "--no-verify", action="store_true",
-        help="Skip post-analysis verification pass",
-    )
-
-
-def build_parser() -> argparse.ArgumentParser:
-    """Build the top-level argument parser with all subcommands."""
-    parser = argparse.ArgumentParser(prog="quodeq")
-    subparsers = parser.add_subparsers(dest="command", required=True)
-
-    dashboard_parser = subparsers.add_parser("dashboard", help="Run the dashboard")
-    dashboard_parser.set_defaults(handler_command="dashboard")
-
-    evaluate_parser = subparsers.add_parser(
-        "evaluate", help="Run evaluation (auto-detects language)"
-    )
-    _add_evaluate_args(evaluate_parser)
-    evaluate_parser.set_defaults(handler_command="evaluate")
-
-    return parser
 
 
 def _resolve_repo(args: argparse.Namespace) -> Path | None:
